Move commandline parsing to separate file.
